Added Further Reading
Hey there - I was looking into Optimism's architecture/rollups and lacked a technical understanding of its distinctiveness, tools, and functions. I found the above links extremely useful; adding them to the page will ease navigation and usability for developers. They strengthen and build upon the content provided in the rest of the page.