Update providers.md
We need to nail down a process for these contributions, this has been our default response. 

Hello, thank you taking the time to write this and I apologize for the delayed response. We're putting a pause on these types of documentation updates until we have a formal process in place. In the meantime please consider filling out this form. We update our ecosystem page weekly and it would be great to have you included.